+++ hc/front/urls.py
@@ -12,6 +12,7 @@ check_urls = [
     path('log/', views.log, name="hc-log"),
     path('status/', views.status_single),
     path('last_ping/', views.ping_details, name="hc-last-ping"),
+    path('transfer/', views.transfer, name="hc-transfer"),
     path('channels/<uuid:channel_code>/enabled', views.switch_channel, name="hc-switch-channel"),
     path('pings/<int:n>/', views.ping_details, name="hc-ping-details"),
 ]